traccia

Function set_hook

pub fn set_hook(hook: Hook)

Sets a hook to be called at the specified point in the logging process.

Multiple hooks can be set.

IMPORTANT: do not use the macros defined in this crate in BeforeLog or AfterLog hooks. This will lead to an infinite loop, as the logging macros trigger the hooks themselves.