traccia

Trait Logger

pub trait Logger: Send + Sync {
    // Required methods
    fn enabled(&self, level: LogLevel) -> bool;
    fn log(&self, record: &Record);
    fn abort(&self);
}

Core trait that defines the logging behavior.

Implementors of this trait handle the actual processing and writing of log records. Custom loggers can be created by implementing this trait.

Required Methods

fn enabled(&self, level: LogLevel) -> bool

Determines if a message with the given log level should be processed.

Arguments

level - The log level to check

Returns

true if messages at this level should be logged, false otherwise

fn log(&self, record: &Record)

Process and output a log record.

Arguments

record - The log record to process

fn abort(&self)

Abort any ongoing logging operations and cleanup resources.

This method is only available when not using the “blocking” feature.