traccia

Trait Style

pub trait Style: Display {
    // Provided methods
    fn bold(&self) -> String { ... }
    fn dim(&self) -> String { ... }
    fn italic(&self) -> String { ... }
    fn underline(&self) -> String { ... }
    fn blink(&self) -> String { ... }
    fn reverse(&self) -> String { ... }
    fn hidden(&self) -> String { ... }
    fn striketrough(&self) -> String { ... }
}

Provided Methods

fn bold(&self) -> String

Applies bold style to the string.

fn dim(&self) -> String

Dims the color of the string.

fn italic(&self) -> String

Applies italic style to the string.

fn underline(&self) -> String

Underlines the string.

fn blink(&self) -> String

Makes the content of the string blink.

fn reverse(&self) -> String

Reverses the foreground and background colors of the string.

fn hidden(&self) -> String

Hides the content of the string with the background color.

fn striketrough(&self) -> String

Strikes through the content of the string.